www/requirements.txt
redis
django
django-celery
